broken Commander into different classes, one per command
