+++ bin/developer-stats.js
@@ -5,11 +5,11 @@ require('dotenv').load();
 module.exports = (function() {
 
     var GithubWrapper = require('../lib/GithubWrapper');
-    var Commander = require('../lib/Commander');
+    var Commander = require('../lib/DeveloperStatsCommander');
 
     var githubWrapper = new GithubWrapper();
 
     var commander = new Commander(githubWrapper);
-    commander.developerStats();
+    commander.start();
 
 }());
