+++ lib/CheckFourEyesCommander.js
@@ -2,9 +2,7 @@ module.exports = (function() {
 
     var inquirer = require('inquirer');
     var Q = require('q');
-    var exec = require('child_process').exec;
     var fs = require('fs');
-    var moment = require('moment');
 
     var CheckFourEyesCommander = function(githubWrapper) {
         this.githubWrapper = githubWrapper;
